core: do not destroy vm on destination host on move to unknown
Since oVirt's early days, we destroy migrating VM on the destination
host when switching it to Unknown status (due to the source
host being non-responsive). Today, this destroy call is redundant:

1. if the migration succeeds before the host becomes responsive again,
the VM switches to Up on the destination host and thus, its run_on_vds
changes to the destination host (and it is destoryed later on the
source host when the latter becomes responsive).

2. if the host becomes responsive again before the migration ends,
we lose the destination host set in the migrating_to_vds field (it is
cleared while switching the VM to Unknown). Today, we manage to cope
with that in order to support migration of hosted engine VMs.

Moreover, the destroy operation only cleaned up the VM from VDSM
but didn't actually destroyed the VM in libvirt in that case until
recently.

Change-Id: I92d949093894ff97ee4f1ab4f3425181cf330649
Related-To: https://bugzilla.redhat.com/1487913
Signed-off-by: Arik Hadas <ahadas@redhat.com>