postgres95: require postgres95 server on plugin-ovirt-engine
For a major upgrade the user is supposed to run only
 yum -y update ovirt-*setup*
so we have to install postgres 95 server in that yum
transaction in order to consume it from engine-setup.
Avoid requiring it in
ovirt-engine-setup-plugin-ovirt-engine-common to avoid
requiring postgres 95 server on hosts where we have just
websocketproxy altough this requires another similar
patch on DWH project.

Change-Id: Id7c97e3c65c44f2e3ec7ff9c26d9735026a3a990
Bug-Url: https://bugzilla.redhat.com/1459134
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>