Add Names To Event Objects
Currently events contain a description that is built a template
and a set of properties added by the backend to the audit log object.
Some applications need to extract certain details from that message.
The most common case is the name of the virtual machine. Sometimes
that information can be extracted from the links that are part of the
event. When this is not possible applications tend to parse the text
message. To avoid this, this patch adds object names references inside
events.

Bug-Url: https://bugzilla.redhat.com/1456711
Change-Id: Ia1d432083bfff5ac923cbee7ec3a3b21e0acca34
Signed-off-by: Ori Liel <oliel@redhat.com>