engine: Delete Cinder disk fails
With unbounded queues the max-threads
is not used and only core-threads are
created. So commands that hold on to the
thread will starve other threads from
executing as in the case of Delete Cinder
Disk command.

Specifying the core-threads to the max-threads
of the pool will create as many threads as
needed upto the number of threads as defined
by core-threads. The unused threads will be
purged after keepalive-time expires cleaning
up the resources used.

Change-Id: I60076ec64b1c430eff44615463a4b0eeeb5f6900
Bug-Url: https://bugzilla.redhat.com/1479869
Signed-off-by: Ravi Nori <rnori@redhat.com>