core: clear migrating_to_vds when migration ends unexpectedly
We typically end VM migration by calling runningSucceeded or runningFailed
of MigrateVmCommand. Other actions cannot be executed on the VM because
the VM is locked. However, in some scenarios, like when the VM is switched
to Unknown, we end the VM migration by calling its reportCompleted method.
In this method, the migrating_to_vds field was not cleared, leading to the
undesired state of having migrating_to_vds!=null and status different than
MigratingFrom. This patch changes the reportCompleted method to clear
that field.

Change-Id: Ib8b4c404eaf5186742d5d6ed3835cd41020e1f0e
Related-To: https://bugzilla.redhat.com/1486543
Signed-off-by: Arik Hadas <ahadas@redhat.com>