core: Filter out memory volumes from unregistered disks.
Unregistered disks are disks which are able to be registered as floating
disks in the oVirt setup.
Memory volumes which are part of a VM's snapshot should not be
considered as floating disks while there is an unregistered VM they are
related to it.

The following patch adds the mapping of memory volumes from the VM's OVF
data as part of the VM's disks, so those will be filtered properly from
the unregistered disks to be import.

Change-Id: I60562245e23208c8820bc77b774e5cc1027f11dd
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>