core: Remove Guid support for flipped bytes
The flipped bytes support in Guid was only useful for the legacy
kerbldap support, which was removed more than two years ago (see
commit e5eac0b).

Thus, there's no longer a reason to keep this dead code. As a side
effect of removing it, Guid(byte[]) was made a tad faster, as there's
no longer a need to dereference the indices array, but that
optimization will probably not be noticeble.

Change-Id: I844596465263437e201e8306def86c0f13e7c738
Signed-off-by: Allon Mureinik <amureini@redhat.com>