core: Check nullness on init of auth-plugin
Under wildfly 11 the plugin handler callback doesn't pass the
'sharedState' initialized. This maybe a wildfly bug or a change in
the default configuration we didn't have till now.

Performing a null-check and not populating
in sharedState in case it is null, makes the management interface accesible
again.

Jboss issue tracker: https://issues.jboss.org/browse/WFLY-9338

Change-Id: I196085d0dc7ffa50703deea541d92a22e94cafb1
Bug-Url: https://bugzilla.redhat.com/1490383
Signed-off-by: Roy Golan <rgolan@redhat.com>