core + webadmin: Manual properties setting for a 'High performance' VM
Manual configuration settings for a new/edited 'High performance' VM/Template/Pool.

This includes displaying a popup dialog for new/edit VM/Pool and for
'new VM from Template' with a list of recommendations for setting the missing
configuration suggestions.

In case the recommended configuration is already set, the related messages won't
be displayed in the popup dialog (or will be partially displayed, depends on
what is missing).
The user will be able to click "cancel" and get back to previous dialog
for fixing the configuration or click "Ok" and ignore.

In case no recommendations exists, the popup won't be displayed.

For 'Edit Pool' action, the dialog is displayed although the Pool can't be
edited. It is displayed just as a reminder that that Pool has a missing configuration
for HP.

For Template and 'edit VM Pools', the dialog is never displayed.

The dialog includes recommendations regarding:
Assigned specific hosts + CPU pinning topology - for both VMs and Pools.
Enable virtual NUMA + set NUMA pinning topology - only for VMs.
Disable KSM (done for the cluster) - for both Vms and Pools.
Enable huge pages - for both Vms and Pools.

In addition to all that, this patch also automatically disabled the
Smartcard device for HP VMs (since it will fail to run without USB).

Change-Id: Ie49ce472b5bc5181de8cae7b17224d88277ab410
Bug-Url: https://bugzilla.redhat.com/1457239
Signed-off-by: Sharon Gratch <sgratch@redhat.com>