restapi: External network provider for hosts
On a host an external network provider could be provision to VMs.
This change enables setting the external network provider during
adding a host, so that oVirt can install the external network
provider during adding the host.
The external network provider is also included in the properties
of the host.

Even if currently only a single external network provider is supported
on a single host the external network provider is encapsulated in a
list, because this way a change in the REST-API is avoided, if multiple
external network providers are supported on a host in future.

Change-Id: I1df1a964b1bbd9c0be3783aa4c365237e9c11471
Bug-Url: https://bugzilla.redhat.com/1495261
Signed-off-by: Dominik Holler <dholler@redhat.com>