core: run add-disk as internal command by transfer disk image
We used to run it as async command using CoCo. But it should actually be
executed as an ordinary internal command. The monitoring is kept the same
- using CoCo's command callback.

This change has two benefits:
1. One less thread from the threadpool is used.
2. More importantly, AddDisk is executed as internal command and thus,
does not try to lock the VM that the disk should be attached to (it
assumes the VM is already locked by the parent command).

Change-Id: I880a5ba9396d9fa2bbaf320691bfde0a6ea0bed6
Signed-off-by: Arik Hadas <ahadas@redhat.com>