core: VmHandler instance methods
This patch changes VmHandler's method that have external dependencies
to be instances method instead of a static methods so thos
dependencies can later be injected.

This change also allows to inject a VmHandler mock to the commands
that use it in their tests instead of having to test the VmHandler's
internal workings in other classes' unit tests.

Change-Id: Id7088fc6d4a319723bf03207c3321f32b11fd466
Signed-off-by: Allon Mureinik <amureini@redhat.com>