restore: delete referenced lun with he-remove-storage-vm
Explicitly perform DeleteLUN when called with
he-remove-storage-vm in order to filter out also the
referenced LUN since only indirectly referenced.

Luns are in luns table, storage connections in
storage_server_connection table and the mapping
between them is in lun_storage_server_connection_map.

We have a FK with ON DELETE CASCADE between
lun_storage_server_connection_map and luns and the
same for lun_storage_server_connection_map and
storage_server_connection.
So either simply deleting from storage_server_connection
or from luns will delete the mapping row in the mapping
table but not in the other table.


Change-Id: I092da786f5e5a9a9f52248bffa02c52953a3b716
Bug-Url: https://bugzilla.redhat.com/1489463
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>