core: Clean up AddVmFromTemplateCommandTest
This patch removes tests that test the inner workings of the
VmHandler and not the AddVmFromTemplateCommand.
As this is a UNIT test for the command, it should not test the
implementation of external dependencies.

The functionality "tested" are just simple pass-through methods, so
there's no real value in recreating them elsewhere.

Change-Id: I50d9227730ae32ddaff725677897b326b1158242
Signed-off-by: Allon Mureinik <amureini@redhat.com>