core: fix enabling fencing from API
When adding/editing a host fencing details from UI both the host
pm_enabled flag and the list of fence agents are set at once and the BLL
handles that.
When you try to do the same from API you need two calls, one for setting
the fence agents (api/hosts/<host-id>/fenceagents) and one to set the
pm_enabled flag for the host (api/hosts/<host-id>), this caused the 2nd
command to fail validation since the original command expects a list of
fence agents to be given.

This patch changes the validation such that if the host update command
was called with an empty or null list of fence agents it will be taken
from the database (if exist)

In addition, in order to align the way that the UX works with the API,
this patch added a validation when a fence agent is removed to reject
the removal if this is the last fence agent and the host pm_enabled flag
is true.

Change-Id: I0c93f39e51e0134b8aabb5b520679b3cb2270b8a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1488755
Signed-off-by: emesika <emesika@redhat.com>