core: ImportVMFromConfigurationCommandTest setup
ImportVMFromConfigurationCommandTest expects a mock OsRepository to
be bound to the SimpleDependencyInjector, but this code is missing.
The test seems to pass in the regular maven build since some test
before it binds the OsRepository (and neglects to clean it up!), but
running this test on its own from an IDE will fail.

This patch adds the required binding and cleanup so that the test can
be run as a standalone test.

Change-Id: Id4e698117caf3c2b4a83e5e5ac6c37ed61fb0a04
Signed-off-by: Allon Mureinik <amureini@redhat.com>