core: disable host fencing in reboot
This patch handles a problem when a host that has a slow start is
fenced.
In that case the host may be rebooted again because the
ServerRebootTimeout is not honored since it is possible for other
threads to call for getting statistics or capabilities for this host.

This patch adds a flag to the VdsManager such that a host is not
handling network exceptions while it is in the ServerRebootTimeout
period.

Change-Id: I28129c49bbdb5409bad67ac11b199ee7075c0b4a
Bug-Url: https://bugzilla.redhat.com/1487981
Bug-Url: https://bugzilla.redhat.com/1477700
Signed-off-by: emesika <emesika@redhat.com>