core: handle controller devices with model='none' in domain xml
in case of a VM set to run without any USB device at all
(device model is set to 'none'), the engine should handle this special
usb device differently.

When such usb+model="none" device appears in domain xml, although this device
is returned without any address, it should still be correlated with the db
devices so it won't be unplugged.

We generalize the usb specific case for fixing the engine handling for any controller
device with model="none" and no given address.

Change-Id: Id26cc9bee887366648dda677fb29c8a5673a8cb6
Bug-Url: https://bugzilla.redhat.com/1457239
Signed-off-by: Sharon Gratch <sgratch@redhat.com>