webadmin: enable high availability for HP VMs
Change the behaviour done in previous patch of always disabling high
availability for a high performance VM.
In this patch we enable setting the VM as highly available for the
pinned hosts (although migration is always disabled for high performance VMs).

This done by canceling the dependency in UI between migration fields to the high
availability field.

Bug-Url: https://bugzilla.redhat.com/1457239

Change-Id: Iba72db9dbf64f8ddb525c8f7b1f0829fb6dc86d0
Signed-off-by: Sharon Gratch <sgratch@redhat.com>