core: errorfix — do not release mac on nic unplug
plug and unplug are called from UI via UpdateVmInterfaceCommand. The
last step is mac pool cleanup, which before this patch assumed, that
mac is being changed. But mac need not to be changed at all and in
that case cleanup should do nothing.

Reason of this bug is then mac not changing during mac unplug, and it
was considered as one to be released.

Change-Id: Ie6ee5e465d86e14da06eb41afc736b5db583f7ca
Bug-Url: https://bugzilla.redhat.com/1492723
Signed-off-by: Martin Mucha <mmucha@redhat.com>