webadmin: generate and package SCSS source maps
To facilitate debugging of CSS/SCSS with browser dev tools,
the ovirt-brand SCSS compiler configuration was updated
to generate CSS source maps inline with the mapping file.
Now the SCSS can be inspected and debugged directly in browser.

The source maps are only generated and packaged when the
engine is build in dev mode (make target install-dev or
option BUILD_DEV=1).

Change-Id: Ica2a8c01edd3ec776c6cdd3040784ebe670cc8fe
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>