webadmin: Reflect cluster's default network provider
On adding a host, initially the network provider is set to the
cluster's default network provider. If the user wants to choose another
network provider, he has to disable a checkbox to be able to select
one of the other network providers.

The user might want to override the default network provider, if she
wants a provider configuration on the host which is different from the
one oVirt is configuring automatically. A plausible motivation to do so
is if the user wants to migrate from one provider to another, but do so
gradually.

Change-Id: Iba890fe8fd432875084f6a1c613351f22f6bcc09
Bug-Url: https://bugzilla.redhat.com/1495261
Signed-off-by: Dominik Holler <dholler@redhat.com>