core: Retry failed live merge commands
If MergeStatusCommand, DestroyImageCommand or DestroyImageCheckCommand
fail, we want to retry executing those commands. At this point of live
merge, if it is an active volume merge, the top volume is marked as
illegal. Failing live merge at one of those commands, ends up with an
illegal volume in the chain and in this case, if the Vm is shutdown, the
user will not be able to start it again.

Change-Id: I6bceba2645084b5e5ba2844ca9af2c4e26e8dfe1
Bug-Url: https://bugzilla.redhat.com/1383301
Bug-Url: https://bugzilla.redhat.com/1467928
Signed-off-by: Ala Hino <ahino@redhat.com>