core: currently only managed host devices are monitored
The current VM devices monitoring is suitable only for managed
host devices, i.e., those that can be correlated with one of
the devices of the host that the engine knows of. This patch
adds proper documentation and checks to prevent exceptions
that may happen when monitoring unmanaged host devices.

Change-Id: I1975fa6c0b723ad8c4a0f65f003fe9600201f2f3
Bug-Url: https://bugzilla.redhat.com/1481007
Signed-off-by: Arik Hadas <ahadas@redhat.com>