core: fix NPE when running vm with mdev device
Recently, a new kind of devices was introduced that although
being host devices, cannot be correlated with any of the devices
of the host that the engine knows of. Those mdev devices thus
break the assumption taken on the engine side that such a
correlation exists and lead to an NPE when running a VM that is
set with these devices. This patch prevents that NPE by ignoring
those devices on run VM.

Change-Id: I0033cc6ccc72f230693bcf6a88a0d641370190ca
Bug-Url: https://bugzilla.redhat.com/1481007
Signed-off-by: Arik Hadas <ahadas@redhat.com>