core: refresh vm devices after skipping unmanaged host device
Previous patch prevented sending unmanaged host devices on run VM.
However, the devices may have no change so their hash remains the
same. In that case, we may be missing those devices that are
re-created by VDSM/libvirt. Therefore, clearing the devices hash
in that case so the VM devices will be refreshed.

Change-Id: Ifd73e2d486ccbe799c426924d2c9a4d45eafc373
Bug-Url: https://bugzilla.redhat.com/1481007
Signed-off-by: Arik Hadas <ahadas@redhat.com>