core: Add support for LUN disks in OVF
Add support for LUN disks as part of the VM's OVF.
LUN disks should be supported in export domain and when importing
an unregistered VM.

Added validation for LUN disks in the import VM command.
On validation if the partial flag will be set to true the VM
will be imported without the LUN disk and an audit log will
be presented to the user.

Change-Id: Id5ba296c7c6c0912a646def9ae6187b2aea0df93
Bug-Url: https://bugzilla.redhat.com/1477505
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>