restapi: Simplify the virtual machine type mappers
Currently the virtual machine type is mapped using the @Mapping
mechanism. That is too heavy for this purpose, and it isn't type safe.
This patch simplifies the code so that it uses plain static methods. In
additions these methods are moved to the VmBaseMapper class, as that is
the only place where they are used.

Change-Id: I15616e593a0a8b66cc231edb9d49f0c3e9e7ba02
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>