webadmin: fix sort snapshot by creation date
Currently, there's no explicit comparator in VmSnapshotListModel,
and sorting is done in 'VmSnapshotListModel -> setItems'.
Hence, since commit I78320a33a27f560f91a956f3bacb4346e0075b48
ensured sorting using comparator with a fallback to ID sorting,
the snapshots list isn't sorted properly (as it should be sorted
by creation_date and not by snapshot_id). Therefore, added a
comparator to VmSnapshotListModel.

Note: the issue affects only 4.1, so pushing this for consistency
with 4.1 code, as in master the snapshot list is implemented by
PatternflyListView - which seems not to be ordered by the
comparator (another bug?).

Change-Id: I918e1bd0d7b0b834f40e5aac824fc3c2d716c1b3
Bug-Url: https://bugzilla.redhat.com/1487291
Signed-off-by: Daniel Erez <derez@redhat.com>