restapi: Remove static map from LinkHelper
The process of link generation in the API relies on static data
of the structure of the API. For historic reasons this data is
currently provided in a hard-coded way by developers. This is
unnecessary since today we have the capabilities to mine this data
automatically. This path removes the manually-maintained map
and replaces it with an automated process

Change-Id: I00da00f298b3b491abdb18023e3b15513a347d0b
Signed-off-by: Ori Liel <oliel@redhat.com>