core: Introduce parameter class for IVTFC
Introduce a new parameter class which will be used for import unregistered
Templates.
The new parameter class will include maps for DR purposes.
Those maps will be used to map any attributes preserved in the entity's OVF
with their correlated attributes in the secondary site.

Here are the mapping which will be supported:
 Cluster
 Dbuser's domain
 Role

Change-Id: I8c16ab493856a8cfc5faa27656b4ded8ef25e6a5
Bug-Url: https://bugzilla.redhat.com/1505436
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>