core: Introduce parameter class for IVFC
Introduce a new parameter class which will be used for import unregistered
VMs.
The new parameter class will include a map for DR purposes.
This map will be used to map any attributes preserved in the entity's OVF
with their correlated attributes in the secondary site.

The maps will be supported for the following attributes:
 Cluster
 External LUN disk
 dbuser's domain
 Role
 Affinity group
 Affinity label

Change-Id: Ia2cb342f5c296c0633967d7d51197fe244c4bbe9
Bug-Url: https://bugzilla.redhat.com/1505436
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>