core: Set imported Memory snapshot image to active
Set memory snapshot images (memory and metadata) to be active
when importing a VM with memory snapshots.

The memory snapshots images (memory and metadata) set as active
images when they create, but after a VM with memory snapshots is
import to the environment, the images are imported as non-active images.

This patch fixes this behavior.

Also add upgrade script to fix the defected existing entities.

Change-Id: Ifb171c28e59be934100cc3d613ebcf33ebe7a91f
Bug-Url: https://bugzilla.redhat.com/1497518
Bug-Url: https://bugzilla.redhat.com/1497512
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>