core: support IO+emulator threads pinning for HP vms
Add Support in VM xml for pinning the IO thread(s) and emulator
thread(s) for a running VM.
This is supported only for High Performance VMs and only if IO thread and NUMA
pinning are enabled for the VM and if CPU pinning topology is set.

The algorithm for finding the physical CPU(s) to pin to is based on the
assumption that it should be pinned to the NUMA where most vCPU(s) are
pinned for performance reasons.

Therefore we search for the NUMA node where most of vCPUs of the
VM are pinned and pin the IO threads (only 1 if set by UI but can be > 1
if set by rest api) and emulator threads to the 1st and 2nd pCPU's of
that NUMA node.

In case of overlapping with pinned vCPU's, a warning is written to the
log.

Change-Id: If12bb0890f758f3cd70c93b89c32fe5555a68c63
Bug-Url: https://bugzilla.redhat.com/1498187
Signed-off-by: Sharon Gratch <sgratch@redhat.com>