core: Retry failed live merge commands
If MergeStatusCommand, DestroyImageCommand or DestroyImageCheckCommand
fail, we want to retry executing those commands. At this point of live
merge, if it is an active volume merge, the top volume is marked as
illegal. Failing live merge at one of those commands, ends up with an
illegal volume in the chain and in this case, if the Vm is shutdown, the
user will not be able to start it again.

Change-Id: I34f367b7162a1ba834a5cf41972cee9011cad6a1
Bug-Url: https://bugzilla.redhat.com/1383301
Bug-Url: https://bugzilla.redhat.com/1467928
Signed-off-by: Ala Hino <ahino@redhat.com>