webadmin: VM table in Host subtab shows pinned VMs
The VM table in Host subtab now shows running VMs on the host
and VMs pinned to the host. A radio button is added to choose
between displaying all, only running, or only pinned VMs.

Here is a pull request for ovirt-site with screenshots
of this change: https://github.com/oVirt/ovirt-site/pull/1264

Change-Id: I8ec15785efc18484feb63ddb0f443ad95e66fa40
Bug-Url: https://bugzilla.redhat.com/1428498
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>