core: Fix RefreshHostDevicesCommandTest
There is no flow in the engine that creates a HostDevice with a null
parent, and the DB even enforces a not null constraint on it.

This patch fixes the test so that the "worse" device has some
fictitious parent name so it represents valid data, and allows for
further refactoring in subsequent patches.

Change-Id: I8b4d9517bf4411e2e5459832e75fe0ff23956bbe
Signed-off-by: Allon Mureinik <amureini@redhat.com>