core: run full host upgrade
This patch add a possibility to run full host update, when user executes
the update of the host and also check-update functionality check for any
update on the host.

This is achieved by new Ansible role[1] added to ovirt-ansible repo.

We update or check for updates for the host  using single playbook called
ovirt-host-upgrade.yml. This playbook just executes the ovirt-host-upgrade
role. The difference between check-updated and update is following:

Check for update:

  - Run playbook with dry-mode (--check) and env variable
    ANSIBLE_STDOUT_CALLBACK=hostupgradeplugin.
  - If any tasks tagged by updatecheck reports that there need to be
    installed/updated package 'hostupgradeplugin.py' we save those packages
    and return list of all those packages.
  - At the engine side we parse the stdout and save all packages that
    needs update

Update:

  - Run playbook without dry-mode (--check) and default stdout callback plugin,
    so log stdout to ovirt-host-mgmt-ansible-*.log file.

[1] https://github.com/oVirt/ovirt-ansible/pull/106

Change-Id: I4ff9d63061ebdc4ec9d98242e9bb281197b5d89d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1380498
Signed-off-by: Ondra Machacek <omachace@redhat.com>