webadmin: Move neutron install warning
Move the warning about the automatic installation of neutron
provider in OpenStack networking specific widget.
This way the warning is only shown, if an external OpenStack
network provider is selected, but not if an oVirt-OVN network
provider is selected.

Additional a AlertPanel is used instead of a custom stylesheet
to show the warning.

Change-Id: I085e686a0e4d83164f164db193524a0f94e49835
Signed-off-by: Dominik Holler <dholler@redhat.com>