core: update welcome page error for missing localized docs
In the case that localize documentation isn't installed or
isn't available, and a user requests docs from the welcome
page, an error is displayed to the user.  The error message
itself has been adjusted so that it can be fully translated
and be grammatically correct.

For example: In German, the localized name of the language is
'Deutch'.  However, when used in a sentence, it is written
as 'deutsche'.  The fix is to change the error text so the
language's name is translated or no longer needed instead
of assuming Java's language display name will be appropriate.

Change-Id: I9076309a1bd466a2bd69f13996aa7eae88e846c7
Bug-Url: https://bugzilla.redhat.com/1356053
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>