webadmin: fix info icon placement on VM edit / system tab
Adjusted the info icon layout on the serial number policy
checkbox on the system tab of VM edit popup so the icon
lines up with the label just like all of the other info
icons.

Added the ability to add a CSS class style directly to
the info icon in an EntityModelWidgetWithInfo widget.

Change-Id: I6640c6d86ebd0af9cb39266f8d070a23cb58c3ee
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>