restapi: Don't load JAX-RS annotations in apidoc
The 'apidoc' application contains a servlet that lives in the 'jaxrs'
module. In order to use it the 'jboss-deployment-structure.xm' file of
this application requires the 'jaxrs' module like this:

  <dependencies>
    <module name="org.ovirt.engine.api.restapi-jaxrs" annotations="true"/>
  </dependencies>

An unexpected side effect of importing the annotations is that the
application server may initialize the JAX-RS application from the
'apidoc' web application instead of from the main web application of the
API. If that happens, some resources, in particular mappers, will not be
loaded, because the code uses the context class loader to load them, and
that context class loader of 'apidoc' doesn't contain them. This doesn't
always happen, as depends in what order the web applications are loaded,
but it looks like in most environments 'apidoc' is loaded before the
main web application. To avoid this issue this patch removes the import
of the annotations, which isn't needed anyhow.

This regression was introduced by the following patch:

  restapi: Integrate API explorer
  https://gerrit.ovirt.org/82401

Change-Id: I4214a9b4fc88e19f091ac1e1ef30198bc4e1289a
Related: https://bugzilla.redhat.com/1404753
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>