packaging: bin: Source engine-prolog.sh directly
Do not source it from generate-pgpass.sh , but from scripts that source
it. Currently only engine-vacuum.sh, but it can be useful for others.

Currently, when some script S sources engine-vacuum.sh, and it sources
engine-prolog.sh using a path relative to $0, "$0" is the path of S, not
of engine-prolog.sh. This breaks if another script wants to source
generate-pgpass.sh, which is not in the same directory.

This is a partial back-port of 8b6e0c9.

Change-Id: I9558d6d4645f352eead58fb6ae5b97edb2b04bb1
Related-To: https://bugzilla.redhat.com/1502592
Signed-off-by: Yedidyah Bar David <didi@redhat.com>