core: Memory hot unplug by decrementing memory size
... to allow memory hot unplug from REST api. It works from both REST
api and Edit Vm dialog.

It tries to pick set of devices total size of which will be the largest
possible smaller than or equal to the size of requested memory
decrement.

This rounding is not applied to next run snapshot value of memory size.

Min memory is decremented as well if possible. Min memory is never
rounded.

Change-Id: I637bffe9a75192f5b9030ae82d6f945619c74ff8
Bug-Url: https://bugzilla.redhat.com/1496382
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>