core: Setting Cinder image's status if not in DB
Before this patch, during snapshot's restore (using REST API) or
snapshot's preview and commit (on UI), a diskImage's status was set to
'Illegal' in case it wasn't a part of the snapshot and so meant to be
deleted. In case of Cinder disks, since their status wasn't set to
Illegal, accessing Cinder disks which are not part of the snapshot
gets a NullPointerException.

Change-Id: I78b92f9c1ebc9c7c1832edd4ed782482bbad9243
Signed-off-by: Shani Leviim <sleviim@redhat.com>