webadmin: Further expand width of QoS popup for Firefox
A previous patch [1] was submitted to expand the width of
the new Storage QoS popup because it was too narrow to
accommodate the width of the fr_FR and es_ES translations.
That patch wasn't tested in Firefox and didn't extend it enough
to accommodate the es_ES translation. This patch corrects that.
The screenshots below [2][3] were taken using Firefox.

[1] https://gerrit.ovirt.org/#/c/81280/
[2] es_ES - https://imgur.com/a/zq778
[3] fr_FR - https://imgur.com/a/juy1y

Change-Id: Iacf38ca36c21055a28ee5ab06c627df4ab25b5ce
Bug-Url: https://bugzilla.redhat.com/1417904
Signed-off-by: Phillip Bailey <phbailey@redhat.com>