engine: Persist speed property for all iface types
The speed property received from the network caps was consumed only for
nic type interfaces.
This patch allows the speed property to be processed (if it exists) for
any interface type, including bonds.

The generalization has been applied on the conversion between the caps
data to engine interface objects.
Other locations (persisting to the DB and reading from it) are already
covered in a generic manner with relation to the speed property.

Change-Id: I4314772ebd9c0c0f59e02e4a93e6a961ead72bd9
Bug-Url: https://bugzilla.redhat.com/1464055
Signed-off-by: Edward Haas <edwardh@redhat.com>