core: Remove check for max images in chain
Commit 4a669e (in oVirt Engine 4.1) added a validation to block the
creation of image chains longer than 90 images due to qemu's
limitation on the path length.

As the underlying cause for these long paths was fixed in VDSM 3.6.7
(see https://bugzilla.redhat.com/1336367 for additional details), and
we no longer support DCs with lower compatibility versions than 3.6,
there's no need for this validation anymore.

This patch removes this now-redundant validation and all its usages,
in order to clean up the code.
As a side bonus, creating a snapshot and running a stateless VM are
now a tad faster, as an one of the DB queries performed for their
validation has been removed.

Change-Id: Id963f3c1dbca515a6f6b4050660f27d64b57ceaa
Signed-off-by: Allon Mureinik <amureini@redhat.com>