webadmin: fix SSH public key field visibility toggle in Safari
The method used to toggle visibility of the password vs SSH
public key field directly changed the visibility style attribute
on the fields. This did not work in Safari on OSX. Changing to use
the method #setVisible() does work.

Event handlers for the password and SSH public key radio buttons
was switched from onFocus to onClick. The onClick event works
more consistently between Chrome, FF and Safari in this case.

Change-Id: I625e739a55b731c9d265ceb65cc8f9ffb2b77241
Bug-Url: https://bugzilla.redhat.com/1480213
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>