core: FenceValidator CDI
FenceValidator is a stateless utility class with external
dependencies, making it a prime candidate for being a @Injected bean.

This patch converts FenceValidator to be a @Singleton, @Injects it to
all the places that need it and @Injects its own dependencies.

Besides cleaning up the code, as a side bonus all the commands that
extend VdsCommand can now reuse the same instance of FenceValidator
instead of each one creating its own.

Coverity-CID: 1382174
Change-Id: Ie84cb9411ae6865222144db2fd5664a65d224991
Signed-off-by: Allon Mureinik <amureini@redhat.com>