core: fail SD deactivation if OVF update failed
When deactivating data storage domain an OVF update occurs.
Currently, if the OVF update fails it is silently ignored, the
storage domain enters to maintenance state and the operation succeeded.

This change prevents a storage domain to be deactivated in case of OVF
update failure, the operation will fail and a proper log message will
appear.

Change-Id: I6deda65dba83e381a99456b399ee962011f5b52a
Bug-Url: https://bugzilla.redhat.com/1321585
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>