engine: SRIOV physical function can have more than one child device
The current code assumed that if a pci device is sriov physical function
it should have at most one child device (net device).
This appears to be wrong (for example in case of iSCSI offload or FCoE,
the pci device has two child devices - 'scsi_host' and 'net_device').
The code was modified to look for the first child device which is 'net'
device, instead of the first child device of any type.

Change-Id: Id569b245b6e5b0647fac0b85788cbb41667cfd12
Bug-Url: https://bugzilla.redhat.com/1506887
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>