core: Remove format validation from ExportVmCommand
Validating that the current format/type of images to be exported is
compatible with the export domain is redundant at best (if both the
source and target domains are file-based domains) and wrong at worst,
since the underlying format/type anyway (see
CopyImageGroupCommand#performStorageOperation).

This patch removes this wrong validation and then the
checkImagesConfiguration method from ImagesHandler, which has no
other usages.

Change-Id: Ie2792375cced0c747bb85ab764aeacdf561e7109
Signed-off-by: Allon Mureinik <amureini@redhat.com>