restapi: Change OpenstackNetworkProvider plugin type to open text
This patch changes OpenstackNetworkProvider to accept the open text
external_plugin_type.
external_plugin_type replaces depracated plugin_type.
external_plugin_type is an open text attribute.
This attribute will allow to choose the correct provider dirver
on the host when an external nic is changed.
The old plugin_type attribute will only be valid for providers
of type NEUTRON and will only be returned when it's value is
equal to OPEN_VSWITCH.
If both 'pluginType' and 'externalPluginType' attributes
are specified during an update, the 'pluginType'
value will be ignored.
For external providers 'plugin_type' will not be shown and will
be ignored during update requests.

Change-Id: Ie2c8f045814d8145baa7d183436151b796f16198
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>