webadmin: don't estimate the total size of a disk
When the total size of the disk is unknown (typically an upload flow via
the REST API), we assume that the client uploads the whole disk and
determine the progress accordingly.
Since in the real world this is not the case, it's better not to assume
anything and to not show a progress at all.

Change-Id: I0cb1c1fb75182e0c3541efe86da3d5d1d3723b6c
Signed-off-by: Idan Shaby <ishaby@redhat.com>