aaa: VM Portal logs out due to missing session refresh
Introduce app-scope a context init parameter
which can be used to configure a scope for
the application that is different from the
scope used to obtain a token from SSO. The
app-scope can be used to determine if the
session needs to be refereshed on each query
performed.

Specifying app-scope of 'ovirt-app-api' will
referesh session on all queries and a scope
of 'ovirt-app-admin' or 'ovirt-app-portal' will
only refresh those queries which have explicity
set the refresh parameter in QueryParametersBase.

Change-Id: I7ddfdcf009d86d12c0db6edc07bd114a7cdcf56c
Bug-Url: https://bugzilla.redhat.com/1508403
Signed-off-by: Ravi Nori <rnori@redhat.com>