[RFC] core: Use osinfo for console target type
The serial console is not working for s390. In order to allow
different console types, the osinfo settings have been extended
with a new device setting "console.target.type" which can have a
value of serial (default), virtio or sclp (used for s390).

This patch doesn't touch the UI, which still says VirtIO Console
for serial consoles.

Change-Id: I4bae536203e49d29987d7901dd6c632c116260c5
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>