engine: Remove Tablet device for High Performance VMs
This is done in order to support running High Performance VMs in
non-headless mode with VNC graphics protocol.

Reason:
When trying to run a High Performance VM as non headless (via regular run or run-once)
and set the graphics protocol to be VNC, the VM failed to start becuase
the Tablet device was added automatically and it requires the USB device
installed. But since USB is not installed for High Performance VM's with VNC
chosen, then the VM failed to run.

Change-Id: I12bbaa3ecfa8564060aac4986ec04a0ca0b49fab
Bug-Url: https://bugzilla.redhat.com/1507847
Signed-off-by: Sharon Gratch <sgratch@redhat.com>